Changed unit-testing to use Django's test.Client
Python's 'requests' module has been replaced by Django's 'test.Client' module.

The Client module uses a temporary test database that is set up before the unit tests, meaning we don't have to rely on hard-coded queries that we expect to be in the database.

As a side-effect, this runs much quicker (< 1 second vs 10 seconds) and cuts down on a lot of excess code lines.

Has been re-tested as well.